requirements.txt
asgiref==3.9.1
certifi==2025.10.5
cfgv==3.4.0
charset-normalizer==3.4.3
distlib==0.4.0
Django==5.1.3
django-cors-headers==4.9.0
django-environ==0.12.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.5.1
drf-yasg==1.21.8
filelock==3.20.0
flake8==7.1.1
gunicorn==21.2.0
identify==2.6.15
idna==3.10
nodeenv==1.9.1
pillow==11.3.0
platformdirs==4.5.0
pre_commit==4.4.0
psycopg2-binary==2.9.11
PyJWT==2.10.1
pytest==8.3.3
pytest-django==4.9.0
PyYAML==6.0.3
requests==2.32.5
sqlparse==0.5.3
urllib3==2.5.0
virtualenv==20.35.4